+++ src/renderer/components/contact/ContactList.tsx
@@ -118,6 +118,7 @@ export function useContactsMap(listFlags: number, queryStr: string) {
   return [contacts, updateContacts] as [typeof contacts, typeof updateContacts]
 }
 
+// The queryStr is trimmed for contact search and for checking validity of the string as an Email address
 export function useContactsNew(listFlags: number, initialQueryStr: string) {
   const [state, setState] = useState<{
     contacts: Type.Contact[]
@@ -131,10 +132,10 @@ export function useContactsNew(listFlags: number, initialQueryStr: string) {
         const contacts = await BackendRemote.rpc.getContacts(
           accountId,
           listFlags,
-          queryStr
+          queryStr.trim()
         )
         const queryStrIsValidEmail = await BackendRemote.rpc.checkEmailValidity(
-          queryStr
+          queryStr.trim()
         )
         setState({ contacts, queryStrIsValidEmail })
       }, 200),